Change package name to javacserver
