+++ make/langtools/tools/javacserver/options/CommandLine.java
@@ -23,7 +23,7 @@
  * questions.
  */
 
-package com.sun.tools.sjavac.options;
+package javacserver.options;
 
 import java.io.IOException;
 import java.io.Reader;
