+++ make/langtools/tools/javacserver/server/Sjavac.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2016, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -23,14 +23,21 @@
  * questions.
  */
 
-package com.sun.tools.sjavac.server;
+package javacserver.server;
+
+import javacserver.Result;
+
 
 /**
+ * Interface of the SjavacImpl, the sjavac client and all wrappers such as
+ * PooledSjavac etc.
+ *
  *  <p><b>This is NOT part of any supported API.
  *  If you write code that depends on this, you do so at your own risk.
  *  This code and its internal interfaces are subject to change or
  *  deletion without notice.</b>
  */
-public interface Terminable {
-    void shutdown(String quitMsg);
+public interface Sjavac {
+    Result compile(String[] args);
+    void shutdown();
 }
