+++ make/langtools/tools/javacserver/server/log/LoggingOutputStream.java
@@ -23,13 +23,13 @@
  * questions.
  */
 
-package com.sun.tools.sjavac.server.log;
+package javacserver.server.log;
 
-import com.sun.tools.sjavac.Log;
 import java.io.ByteArrayOutputStream;
 import java.io.FilterOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import javacserver.Log;
 
 public class LoggingOutputStream extends FilterOutputStream {
